Move Useless Exceptions out of Client
There are work server specific exceptions
living in the mirrulations-client package.
These are being imported in
the mirrulations-work-server package.
These exceptions do not belong in the client.
They should be dropped when the work server is
dropped.
